Made example notebook for Karanovic solution
Example compares with Domenico solution and with results from BIOSCREEN-AT

Additionally: fixed bug where source zone concentration/boundary would be seen as an integer array if input array only contained integers.

+++ mibitrans/data/check_input.py
@@ -94,9 +94,9 @@ def validate_source_zones(boundary, concentration):
         boundary = np.array(boundary)
 
     if isinstance(concentration, (float, int, np.floating, np.integer)):
-        concentration = np.array([concentration])
+        concentration = np.array([concentration], dtype=float)
     else:
-        concentration = np.array(concentration)
+        concentration = np.array(concentration, dtype=float)
 
     # Each given source zone boundary should have a given concentration, and vice versa
     if boundary.shape != concentration.shape:
